NODE-1860: Refactoring to the last version of WEP

+++ node-it/src/test/scala/com/wavesplatform/it/api/model.scala
@@ -207,7 +207,7 @@ case class Block(signature: String,
                  transactions: Seq[Transaction],
                  fee: Long,
                  features: Option[Seq[Short]],
-                 reward: Option[Byte])
+                 reward: Option[Long])
 object Block {
   implicit val blockFormat: Format[Block] = Json.format
 }
@@ -219,7 +219,7 @@ case class BlockHeaders(signature: String,
                         transactionCount: Int,
                         blocksize: Int,
                         features: Option[Set[Short]],
-                        reward: Option[Byte],
+                        reward: Option[Long],
                         totalFee: Long)
 object BlockHeaders {
   implicit val blockHeadersFormat: Format[BlockHeaders] = Json.format
